progress

+++ CoreEngines/src/HealthComponent.h
@@ -6,7 +6,7 @@ namespace nb
 	class HealthComponent : public Component
 	{
 		const int m_maxHealth;
-		int m_health;
+		float m_health;
 
 	public:
 		HealthComponent( int maxHealth,
@@ -16,9 +16,9 @@ namespace nb
 
 		virtual void destroy() override;
 
-		void damage( unsigned int amount );
+		void damage( float amount );
 
-		void heal( unsigned int amount );
+		void heal( float amount );
 
 		int getHealth()const;
 		int getMaxHealth()const;
@@ -26,8 +26,8 @@ namespace nb
 		bool isDead()const;
 
 		Signal<void( const HealthComponent*const, int )> s_onChange;
-		Signal<void( const HealthComponent*const, unsigned int )> s_onDamage;
-		Signal<void( const HealthComponent*const, unsigned int )> s_onHeal;
+		Signal<void( const HealthComponent*const, float )> s_onDamage;
+		Signal<void( const HealthComponent*const, float )> s_onHeal;
 		Signal<void( const HealthComponent*const )> s_onDie;
 		Signal<void( const HealthComponent*const )> s_onResurrect;
 	};
